pyproject.toml
[project]
name = "project-name"
version = "0.1.0"
description = "프로젝트 설명"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0.0",
    "tqdm>=4.60.0",
    "transformers>=4.30.0",
    "pyyaml>=6.0",
    "mcp>=1.23.1",
    "pynvml>=13.0.1",
    "wandb>=0.24.0",
    "pandas>=2.3.3",
    "matplotlib>=3.10.8",
    "pytest>=9.0.2",
    "nvidia-ml-py>=13.590.44",
]

[project.optional-dependencies]
train = [
    "accelerate>=0.20.0",
    "datasets>=2.10.0",
    "tensorboard>=2.10.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
tools = [
    "skill-seekers>=2.5.0",
]

[tool.ruff]
# Enable pycodestyle (`E`), Pyflakes (`F`), and isort (`I`) codes by default.
select = ["E", "F", "I", "W", "B"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

line-length = 88
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]

# [tool.uv]
# pip = { index-url = "https://pypi.org/simple" }
